+++ administrator/components/com_content/src/Field/Modal/ArticleField.php
@@ -11,13 +11,9 @@
 namespace Joomla\Component\Content\Administrator\Field\Modal;
 
 use Joomla\CMS\Factory;
-use Joomla\CMS\Form\FormField;
-use Joomla\CMS\HTML\HTMLHelper;
-use Joomla\CMS\Language\LanguageHelper;
+use Joomla\CMS\Form\Field\ModalField;
 use Joomla\CMS\Language\Text;
 use Joomla\CMS\Session\Session;
-use Joomla\Database\ParameterType;
-
 // phpcs:disable PSR1.Files.SideEffects
 \defined('_JEXEC') or die;
 // phpcs:enable PSR1.Files.SideEffects
@@ -27,7 +23,7 @@ use Joomla\Database\ParameterType;
  *
  * @since  1.6
  */
-class ArticleField extends FormField
+class ArticleField extends ModalField
 {
     /**
      * The form field type.
@@ -35,7 +31,7 @@ class ArticleField extends FormField
      * @var    string
      * @since  1.6
      */
-    protected $type = 'Modal_Article';
+    protected $type = 'Article';
 
     /**
      * Method to get the field input markup.
@@ -46,53 +42,21 @@ class ArticleField extends FormField
      */
     protected function getInput()
     {
-        $allowNew       = ((string) $this->element['new'] == 'true');
-        $allowEdit      = ((string) $this->element['edit'] == 'true');
-        $allowClear     = ((string) $this->element['clear'] != 'false');
-        $allowSelect    = ((string) $this->element['select'] != 'false');
-        $allowPropagate = ((string) $this->element['propagate'] == 'true');
-
-        $languages = LanguageHelper::getContentLanguages(array(0, 1), false);
-
-        // Load language
+        // Load language.
         Factory::getLanguage()->load('com_content', JPATH_ADMINISTRATOR);
 
-        // The active article id field.
-        $value = (int) $this->value ?: '';
+        return parent::getInput();
+    }
 
-        // Create the modal id.
-        $modalId = 'Article_' . $this->id;
-
-        /** @var \Joomla\CMS\WebAsset\WebAssetManager $wa */
-        $wa = Factory::getApplication()->getDocument()->getWebAssetManager();
-
-        // Add the modal field script to the document head.
-        $wa->useScript('field.modal-fields');
-
-        // Script to proxy the select modal function to the modal-fields.js file.
-        if ($allowSelect) {
-            static $scriptSelect = null;
-
-            if (is_null($scriptSelect)) {
-                $scriptSelect = array();
-            }
-
-            if (!isset($scriptSelect[$this->id])) {
-                $wa->addInlineScript(
-                    "
-				window.jSelectArticle_" . $this->id . " = function (id, title, catid, object, url, language) {
-					window.processModalSelect('Article', '" . $this->id . "', id, title, catid, object, url, language);
-				}",
-                    [],
-                    ['type' => 'module']
-                );
-
-                Text::script('JGLOBAL_ASSOCIATIONS_PROPAGATE_FAILED');
-
-                $scriptSelect[$this->id] = true;
-            }
-        }
-
+    /**
+     * Method to get the urls used by this field.
+     *
+     * @return  array
+     *
+     * @since   __DEPLOY_VERSION__
+     */
+    protected function getUrls()
+    {
         // Setup variables for display.
         $linkArticles = 'index.php?option=com_content&amp;view=articles&amp;layout=modal&amp;tmpl=component&amp;' . Session::getFormToken() . '=1';
         $linkArticle  = 'index.php?option=com_content&amp;view=article&amp;layout=modal&amp;tmpl=component&amp;' . Session::getFormToken() . '=1';
@@ -100,14 +64,46 @@ class ArticleField extends FormField
         if (isset($this->element['language'])) {
             $linkArticles .= '&amp;forcedLanguage=' . $this->element['language'];
             $linkArticle  .= '&amp;forcedLanguage=' . $this->element['language'];
-            $modalTitle    = Text::_('COM_CONTENT_SELECT_AN_ARTICLE') . ' &#8212; ' . $this->element['label'];
-        } else {
-            $modalTitle    = Text::_('COM_CONTENT_SELECT_AN_ARTICLE');
         }
 
-        $urlSelect = $linkArticles . '&amp;function=jSelectArticle_' . $this->id;
-        $urlEdit   = $linkArticle . '&amp;task=article.edit&amp;id=\' + document.getElementById(&quot;' . $this->id . '_id&quot;).value + \'';
-        $urlNew    = $linkArticle . '&amp;task=article.add';
+        return [
+            'select' => $linkArticles . '&amp;function=jModalSelect',
+            // The edit url needs the id of the selected item. It will be inserted dynamically in place of ${0}
+            'edit'   => $linkArticle . '&amp;task=article.edit&amp;id=${0}',
+            'new'    => $linkArticle . '&amp;task=article.add',
+        ];
+    }
+
+    /**
+     * Method to get the strings used by this field.
+     *
+     * @return  array
+     *
+     * @since   __DEPLOY_VERSION__
+     */
+    protected function getText()
+    {
+        $selectTitle = Text::_('COM_CONTENT_CHANGE_ARTICLE') . (isset($this->element['language']) ? ' &#8212; ' . $this->element['label'] : '');
+
+        return array_merge(parent::getText(), [
+            'title-select' => $selectTitle,
+            'title-new'    => Text::_('COM_CONTENT_NEW_ARTICLE'),
+            'title-edit'   => Text::_('COM_CONTENT_EDIT_ARTICLE'),
+            'placeholder'  => Text::_('COM_CONTENT_CHANGE_ARTICLE'),
+        ]);
+    }
+
+    /**
+     * Method to get the title of the currently selected item.
+     *
+     * @return  string
+     *
+     * @since   __DEPLOY_VERSION__
+     */
+    protected function getSelectedTitle()
+    {
+        // The active article id field.
+        $value = (int) $this->value > 0 ? (int) $this->value : '';
 
         if ($value) {
             $db    = $this->getDatabase();
@@ -123,181 +119,23 @@ class ArticleField extends FormField
             } catch (\RuntimeException $e) {
                 Factory::getApplication()->enqueueMessage($e->getMessage(), 'error');
             }
-            if (empty($title)) {
-                $value = '';
-            }
         }
 
-        $title = empty($title) ? Text::_('COM_CONTENT_SELECT_AN_ARTICLE') : htmlspecialchars($title, ENT_QUOTES, 'UTF-8');
-
-        // The current article display field.
-        $html  = '';
-
-        if ($allowSelect || $allowNew || $allowEdit || $allowClear) {
-            $html .= '<span class="input-group">';
-        }
-
-        $html .= '<input class="form-control" id="' . $this->id . '_name" type="text" value="' . $title . '" readonly size="35">';
-
-        // Select article button
-        if ($allowSelect) {
-            $html .= '<button'
-                . ' class="btn btn-primary' . ($value ? ' hidden' : '') . '"'
-                . ' id="' . $this->id . '_select"'
-                . ' data-bs-toggle="modal"'
-                . ' type="button"'
-                . ' data-bs-target="#ModalSelect' . $modalId . '">'
-                . '<span class="icon-file" aria-hidden="true"></span> ' . Text::_('JSELECT')
-                . '</button>';
-        }
-
-        // New article button
-        if ($allowNew) {
-            $html .= '<button'
-                . ' class="btn btn-secondary' . ($value ? ' hidden' : '') . '"'
-                . ' id="' . $this->id . '_new"'
-                . ' data-bs-toggle="modal"'
-                . ' type="button"'
-                . ' data-bs-target="#ModalNew' . $modalId . '">'
-                . '<span class="icon-plus" aria-hidden="true"></span> ' . Text::_('JACTION_CREATE')
-                . '</button>';
-        }
-
-        // Edit article button
-        if ($allowEdit) {
-            $html .= '<button'
-                . ' class="btn btn-primary' . ($value ? '' : ' hidden') . '"'
-                . ' id="' . $this->id . '_edit"'
-                . ' data-bs-toggle="modal"'
-                . ' type="button"'
-                . ' data-bs-target="#ModalEdit' . $modalId . '">'
-                . '<span class="icon-pen-square" aria-hidden="true"></span> ' . Text::_('JACTION_EDIT')
-                . '</button>';
-        }
-
-        // Clear article button
-        if ($allowClear) {
-            $html .= '<button'
-                . ' class="btn btn-secondary' . ($value ? '' : ' hidden') . '"'
-                . ' id="' . $this->id . '_clear"'
-                . ' type="button"'
-                . ' onclick="window.processModalParent(\'' . $this->id . '\'); return false;">'
-                . '<span class="icon-times" aria-hidden="true"></span> ' . Text::_('JCLEAR')
-                . '</button>';
-        }
-
-        // Propagate article button
-        if ($allowPropagate && count($languages) > 2) {
-            // Strip off language tag at the end
-            $tagLength = (int) strlen($this->element['language']);
-            $callbackFunctionStem = substr("jSelectArticle_" . $this->id, 0, -$tagLength);
-
-            $html .= '<button'
-            . ' class="btn btn-primary' . ($value ? '' : ' hidden') . '"'
-            . ' type="button"'
-            . ' id="' . $this->id . '_propagate"'
-            . ' title="' . Text::_('JGLOBAL_ASSOCIATIONS_PROPAGATE_TIP') . '"'
-            . ' onclick="Joomla.propagateAssociation(\'' . $this->id . '\', \'' . $callbackFunctionStem . '\');">'
-            . '<span class="icon-sync" aria-hidden="true"></span> ' . Text::_('JGLOBAL_ASSOCIATIONS_PROPAGATE_BUTTON')
-            . '</button>';
-        }
-
-        if ($allowSelect || $allowNew || $allowEdit || $allowClear) {
-            $html .= '</span>';
-        }
-
-        // Select article modal
-        if ($allowSelect) {
-            $html .= HTMLHelper::_(
-                'bootstrap.renderModal',
-                'ModalSelect' . $modalId,
-                array(
-                    'title'       => $modalTitle,
-                    'url'         => $urlSelect,
-                    'height'      => '400px',
-                    'width'       => '800px',
-                    'bodyHeight'  => 70,
-                    'modalWidth'  => 80,
-                    'footer'      => '<button type="button" class="btn btn-secondary" data-bs-dismiss="modal">'
-                                        . Text::_('JLIB_HTML_BEHAVIOR_CLOSE') . '</button>',
-                )
-            );
-        }
-
-        // New article modal
-        if ($allowNew) {
-            $html .= HTMLHelper::_(
-                'bootstrap.renderModal',
-                'ModalNew' . $modalId,
-                array(
-                    'title'       => Text::_('COM_CONTENT_NEW_ARTICLE'),
-                    'backdrop'    => 'static',
-                    'keyboard'    => false,
-                    'closeButton' => false,
-                    'url'         => $urlNew,
-                    'height'      => '400px',
-                    'width'       => '800px',
-                    'bodyHeight'  => 70,
-                    'modalWidth'  => 80,
-                    'footer'      => '<button type="button" class="btn btn-secondary"'
-                            . ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'add\', \'article\', \'cancel\', \'item-form\'); return false;">'
-                            . Text::_('JLIB_HTML_BEHAVIOR_CLOSE') . '</button>'
-                            . '<button type="button" class="btn btn-primary"'
-                            . ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'add\', \'article\', \'save\', \'item-form\'); return false;">'
-                            . Text::_('JSAVE') . '</button>'
-                            . '<button type="button" class="btn btn-success"'
-                            . ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'add\', \'article\', \'apply\', \'item-form\'); return false;">'
-                            . Text::_('JAPPLY') . '</button>',
-                )
-            );
-        }
-
-        // Edit article modal
-        if ($allowEdit) {
-            $html .= HTMLHelper::_(
-                'bootstrap.renderModal',
-                'ModalEdit' . $modalId,
-                array(
-                    'title'       => Text::_('COM_CONTENT_EDIT_ARTICLE'),
-                    'backdrop'    => 'static',
-                    'keyboard'    => false,
-                    'closeButton' => false,
-                    'url'         => $urlEdit,
-                    'height'      => '400px',
-                    'width'       => '800px',
-                    'bodyHeight'  => 70,
-                    'modalWidth'  => 80,
-                    'footer'      => '<button type="button" class="btn btn-secondary"'
-                            . ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'edit\', \'article\', \'cancel\', \'item-form\'); return false;">'
-                            . Text::_('JLIB_HTML_BEHAVIOR_CLOSE') . '</button>'
-                            . '<button type="button" class="btn btn-primary"'
-                            . ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'edit\', \'article\', \'save\', \'item-form\'); return false;">'
-                            . Text::_('JSAVE') . '</button>'
-                            . '<button type="button" class="btn btn-success"'
-                            . ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'edit\', \'article\', \'apply\', \'item-form\'); return false;">'
-                            . Text::_('JAPPLY') . '</button>',
-                )
-            );
-        }
-
-        // Note: class='required' for client side validation.
-        $class = $this->required ? ' class="required modal-value"' : '';
-
-        $html .= '<input type="hidden" id="' . $this->id . '_id" ' . $class . ' data-required="' . (int) $this->required . '" name="' . $this->name
-            . '" data-text="' . htmlspecialchars(Text::_('COM_CONTENT_SELECT_AN_ARTICLE'), ENT_COMPAT, 'UTF-8') . '" value="' . $value . '">';
-
-        return $html;
+        return empty($title) ? Text::_('COM_CONTENT_SELECT_AN_ARTICLE') : htmlspecialchars($title, ENT_QUOTES, 'UTF-8');
     }
 
     /**
-     * Method to get the field label markup.
+     * Method to get attributes for the html tag.
      *
-     * @return  string  The field label markup.
+     * @return  string
      *
-     * @since   3.4
+     * @since   __DEPLOY_VERSION__
      */
-    protected function getLabel()
+    protected function getAttributes()
     {
-        return str_replace($this->id, $this->id . '_name', parent::getLabel());
+        return [
+            'item-type' => 'article',
+            'form-id'   => 'item-form',
+        ];
     }
 }
